requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/ML Dependencies
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10
openai
chromadb==0.4.18
sentence-transformers==2.2.2

# File Processing
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.1.4

# AWS
boto3==1.34.0
botocore==1.34.0

# Monitoring & Logging
sentry-sdk[fastapi]==1.38.0
wandb==0.16.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
